// src/data/gymMapEntitiesGenerated.ts
export const GYM_MAP_ENTITIES: Record<string, TrainerPoint[]> = {
  "dewfordtown-gym": [
    {
      id: "gym-e5",
      name: "Gym Leader Brawly",
      category: "trainer",
      trainerClass: "Gym Leader",
      trainerName: "Brawly",
      trainerId: "TRAINER_BRAWLY_1",
      x: 25,
      y: 14.29,
      graphicsId: "OBJ_EVENT_GFX_BRAWLY",
      spriteSheet: "sprites/trainers/brawly.png",
      spriteWidth: 16,
      spriteHeight: 32,
      spriteFrame: 0,
      
      script: "DewfordTown_Gym_EventScript_Brawly",
      trainerType: "TRAINER_TYPE_NONE",
    },
// ...
};